Add config for uniform noise in period region
Added functionality to be able to control a specific region of the
period of the wave form and add uniform random noise (controllable
bounds) to the wave. This is just an example of programming some artifact
into the wave form that can be used to train models in anomaly
detection.